+++ bin/FilterSeq.py
@@ -4,10 +4,12 @@ Filters sequences in FASTA/FASTQ files
 """
 # Info
 __author__ = 'Jason Anthony Vander Heiden'
-from presto import (__version__, __date__)
+from presto import __version__, __date__
 
 # Imports
-import os, sys, textwrap
+import os
+import sys
+import textwrap
 from argparse import ArgumentParser
 from collections import OrderedDict
 from itertools import groupby
@@ -16,8 +18,8 @@ from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
 
 # Presto imports
-from presto.Core import default_min_qual, default_out_args, default_missing_chars
-from presto.Core import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.Defaults import default_missing_chars, default_min_qual, default_out_args
+from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
 from presto.IO import getFileType, printLog
 from presto.Multiprocessing import SeqResult, manageProcesses, feedSeqQueue, \
                                    processSeqQueue, collectSeqQueue