+++ bin/MaskPrimers.py
@@ -4,18 +4,20 @@ Removes primers and annotates sequences with primer and barcode identifiers
 """
 # Info
 __author__ = 'Jason Anthony Vander Heiden'
-from presto import (__version__, __date__)
+from presto import __version__, __date__
 
 # Imports
-import os, sys, textwrap
+import os
+import sys
+import textwrap
 from argparse import ArgumentParser
 from collections import OrderedDict
 from itertools import izip
 from Bio import pairwise2
 
 # Presto imports
-from presto.Core import default_delimiter, default_out_args
-from presto.Core import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.Defaults import default_delimiter, default_out_args
+from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
 from presto.Annotation import parseAnnotation, flattenAnnotation, mergeAnnotation
 from presto.Sequence import compilePrimers, getScoreDict, reverseComplement
 from presto.IO import readPrimerFile, printLog