+++ bin/PairSeq.py
@@ -4,18 +4,19 @@ Sorts and matches sequence records with matching coordinates across files
 """
 # Info
 __author__ = 'Jason Anthony Vander Heiden'
-from presto import (__version__, __date__)
+from presto import __version__, __date__
 
 # Imports
-import os, sys, textwrap
+import os
+import textwrap
 from argparse import ArgumentParser
 from collections import OrderedDict
 from time import time
 from Bio import SeqIO
 
 # Presto imports
-from presto.Core import default_coord_choices, default_coord_type, default_out_args
-from presto.Core import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.Defaults import default_coord_choices, default_coord_type, default_out_args
+from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
 from presto.Annotation import parseAnnotation, flattenAnnotation, mergeAnnotation, \
                               getCoordKey
 from presto.IO import getFileType, readSeqFile, countSeqFile, getOutputHandle, \