+++ bin/SplitSeq.py
@@ -4,18 +4,21 @@ Sorts, samples and splits FASTA/FASTQ sequence files
 """
 # Info
 __author__ = 'Jason Anthony Vander Heiden'
-from presto import (__version__, __date__)
+from presto import __version__, __date__
 
 # Imports
-import os, random, sys, textwrap
+import os
+import random
+import sys
+import textwrap
 from argparse import ArgumentParser
 from collections import OrderedDict
 from time import time
 from Bio import SeqIO
 
 # Presto imports
-from presto.Core import default_coord_choices, default_coord_type, default_out_args
-from presto.Core import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.Defaults import default_coord_choices, default_coord_type, default_out_args
+from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
 from presto.Sequence import subsetSeqIndex
 from presto.Annotation import parseAnnotation, getAnnotationValues, getCoordKey
 from presto.IO import getFileType, readSeqFile, countSeqFile, getOutputHandle, \