+++ presto/__init__.py
@@ -1,4 +1,11 @@
-from ._version import __version__
-from ._version import __date__
-from ._version import __copyright__
-from ._version import __license__
+# Set package info
+from .Version import __author__
+from .Version import __version__
+from .Version import __date__
+from .Version import __copyright__
+from .Version import __license__
+
+# Set package level imports
+__all__ = ['Defaults']
+from Defaults import *
+